+++ .devcontainer/Dockerfile
@@ -1,0 +1,4 @@
+FROM pokesay-go:latest
+
+RUN go get -v github.com/ramya-rao-a/go-outline \
+    && go get -v golang.org/x/tools/gopls
