+++ go.mod
@@ -2,4 +2,4 @@ module github.com/tmck-code/pokesay-go
 
 go 1.17
 
-require github.com/mitchellh/go-wordwrap v1.0.1 // indirect
+require github.com/mitchellh/go-wordwrap v1.0.1
